requirements.txt
python-escpos==3.1
Pillow==10.4.0
backports.zoneinfo==0.2.1; python_version < "3.9"
